replace functions to utils
